Use uint8 instead of uint256 for interestRate
Unfortunately Vyper is not yet going to pack the struct to save space
for this. But it still seems valuable to avoid ludicrous values; perhaps
an explicit check with a lower max is still worthwhile too (in which
case uint256 might be better again to avoid the superfluous Vyper
checks).