Start removing lender IDs
We can identify lenders simply by their address - nothing rides on their
identity. Previously we cared about lender id to ensure poolIds were
unique, but this seems unnecessary (if the same lender wants to create
many pools with the same parameters that seems fine - they are not
incentivised to do so since it just fragments liquidity). Now we use a
simple nonce for pool identifiers directly.